requirements.txt
pinecone
google-genai
dotenv
pandas
openpyxl
xlrd
praw
google-api-python-client
pandas
matplotlib
numpy
python-dotenv
google-generativeai